backend/requirements.txt
Flask==2.1.2
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-CORS==3.0.10
pymongo==4.1.1
python-dotenv==0.21.0
